% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find_model_data}
\alias{find_model_data}
\title{Find and process a model file.}
\usage{
find_model_data(directory, file)
}
\arguments{
\item{directory}{start point for the (ascending) search.}

\item{file}{name to search for.}
}
\value{
name list with parsed data from the file.
}
\description{
Find and process a model file.
}

