% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_paths}
\alias{get_paths}
\title{Reads a simple tag:value file format.}
\usage{
get_paths(file = "file-data.txt")
}
\arguments{
\item{file}{source file.}
}
\value{
named vector with values as values and tags as names.
}
\description{
Reads a simple tag:value file format.
}

