% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan-names.R
\name{grab_control_lines}
\alias{grab_control_lines}
\title{Grab the lines of a Stan .csv file with the control parameters.}
\usage{
grab_control_lines(lines)
}
\arguments{
\item{lines}{lines from the Stan .csv file, need only enough to get past the recorded meta-data.}
}
\value{
only the lines with the A=B format.
}
\description{
Grab the lines of a Stan .csv file with the control parameters.
}

