% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan-names.R
\name{grab_parameter_names}
\alias{grab_parameter_names}
\title{Use Stan .csv column named parameter from the Stan model.}
\usage{
grab_parameter_names(columns)
}
\arguments{
\item{columns}{.csv column names with name.i.j.... format.}
}
\value{
A vector of unique named parameters.  Should match parameters
named in the Stan model, plus internal (xxx__) parameters.
}
\description{
Use Stan .csv column named parameter from the Stan model.
}

