% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan-names.R
\name{grab_parameter_vector}
\alias{grab_parameter_vector}
\title{Grab the line of a Stan .csv file with the column names in it, starts
with "lp__".}
\usage{
grab_parameter_vector(lines)
}
\arguments{
\item{lines}{lines from the Stan .csv file, need only enough to get past the recorded meta-data.}
}
\value{
vector of all parameter names, with whitespace trimmed.
}
\description{
Grab the line of a Stan .csv file with the column names in it, starts
with "lp__".
}

