% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{map_block-class}
\alias{map_block}
\alias{map_block-class}
\title{A reference class for constructing the model matrix for either a
spline (when a distance/weight function is passed as the helper) or
an arbitrary factor mapping (when the mapping function is passed as
the weight_helper).  In either case, the 'covariate' can be a list
s.t. weight_helper can access multiple covariates.}
\description{
A reference class for constructing the model matrix for either a
spline (when a distance/weight function is passed as the helper) or
an arbitrary factor mapping (when the mapping function is passed as
the weight_helper).  In either case, the 'covariate' can be a list
s.t. weight_helper can access multiple covariates.
}
\section{Methods}{

\describe{
\item{\code{initialize(name, covariate = list(), validation = NULL)}}{Initialize the base class with a name and validation function.}

\item{\code{make_block()}}{For the base class this is a placeholder which throws an error.
In all other subclasses this function arranges the covariates into the
class-specific model matrix.  The name of the model}
}}

