% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{model_block-class}
\alias{model_block}
\alias{model_block-class}
\title{A reference base class for classes which construct parts of a model
matrix.  Each part is a set of columns referring to one effect.
The goal is to focus on construction of model matrices, delegate when
possible to model.matrix, and avoid dealing with 'formula'-style
interfaces or additional issues (constraints, model types, etc...)}
\description{
A reference base class for classes which construct parts of a model
matrix.  Each part is a set of columns referring to one effect.
The goal is to focus on construction of model matrices, delegate when
possible to model.matrix, and avoid dealing with 'formula'-style
interfaces or additional issues (constraints, model types, etc...)
}
\section{Fields}{

\describe{
\item{\code{effect_name__}}{A character string with the overall name of the
effect.}

\item{\code{input_check__}}{A function for validating covariate values.}

\item{\code{X__}}{Matrix holding the block in question.}
}}
\section{Methods}{

\describe{
\item{\code{initialize(name, covariate = list(), validation = NULL)}}{Initialize the base class with a name and validation function.}

\item{\code{make_block()}}{For the base class this is a placeholder which throws an error.
In all other subclasses this function arranges the covariates into the
class-specific model matrix.  The name of the model}
}}

