% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multifile-accessors-methods.R
\docType{methods}
\name{names,stan_commander-method}
\alias{names,stan_commander-method}
\title{Get names of model parameters contained in stan_commander object.}
\usage{
\S4method{names}{stan_commander}(x)
}
\arguments{
\item{x}{stan_commander object.}
}
\value{
unique parameters contained in the object.
}
\description{
Get names of model parameters contained in stan_commander object.
}

