% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{offset_block-class}
\alias{offset_block}
\alias{offset_block-class}
\title{A reference class for constructing the (simple!) model matrix for a
factor, this preduces a parameterization assuming some other
intercept.}
\description{
A reference class for constructing the (simple!) model matrix for a
factor, this preduces a parameterization assuming some other
intercept.
}
\section{Methods}{

\describe{
\item{\code{initialize(name, covariate, formula = NULL, validation = NULL)}}{Builds on initialization for base class, but only by calling method to construct the block.}

\item{\code{make_block(covariate)}}{Construct model matrix, delegate to model.matrix.}
}}

