% Generated by roxygen2 (4.0.2): do not edit by hand
\name{path_split}
\alias{path_split}
\title{Normalizes a path and splits it on '/'.}
\usage{
path_split(path)
}
\arguments{
\item{path}{non-normalized paths.}
}
\value{
list of normalized and split path parts.
}
\description{
Normalizes a path and splits it on '/'.
}

