% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_stan_file}
\alias{read_stan_file}
\title{Properly read a stan file, avoiding factors and comments, including
reading column names from a header.}
\usage{
read_stan_file(file)
}
\arguments{
\item{file}{An output file produced by CmdStan.}
}
\value{
A matrix with samples or optima.
}
\description{
Properly read a stan file, avoiding factors and comments, including
reading column names from a header.
}

