% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{stan_commander-class}
\alias{stan_commander}
\alias{stan_commander-class}
\title{A reference class allows access to output from multiple CmdStan .csv
files efficiently.}
\description{
A reference class allows access to output from multiple CmdStan .csv
files efficiently.
}
\section{Fields}{

\describe{
\item{\code{paths}}{paths to .csv output from optimize and/or sample.}

\item{\code{type}}{vector indicating result type ("optimize" or "sample").}

\item{\code{id}}{vector indicating the try/chain number.}

\item{\code{model_parameters}}{names of model parameters (per-estimation)}

\item{\code{internal_parameters}}{names of internal parameters (per-estimation)}

\item{\code{meta}}{other data on output provided by parse_parameters function.}

\item{\code{dimensions}}{list (one entry per named parameter) listing dimensions of said parameter.}

\item{\code{estimates}}{list (one entry per estimate source) of matrices with sample/optim solution estimates.}
}}
\section{Methods}{

\describe{
\item{\code{check_hashes()}}{If hashes are stale, refresh the object and update hashes.}

\item{\code{do_hashes()}}{Calculates current hashes of files in 'paths' field.}

\item{\code{do_meta()}}{Parse file comments/header.}

\item{\code{do_reload()}}{Parse file estimates.}

\item{\code{get_parameter(x, ...)}}{Implementation for the subset operator '['... used directly or via an S4 method.}

\item{\code{initialize(paths)}}{Loads data from files, and fill fields.}

\item{\code{make_name(name, indexes)}}{Paste together parameter name with indexes to generate csv column name.}

\item{\code{refresh()}}{Reload object data from files.}

\item{\code{stale_hashes()}}{Test if the current exposed hashes ('hashes') are out of sync with current files.}
}}

