% Generated by roxygen2 (4.0.2): do not edit by hand
\name{string_masks}
\alias{string_masks}
\title{Takes strings, check if they match patterns, and construct expressions according to a string.}
\usage{
string_masks(x, patterns, expr = NULL, all_strings = TRUE)
}
\arguments{
\item{x}{A vector of strings which masks will be created for.}

\item{patterns}{A list of patterns (for grepl) which the strings (might) match.}

\item{expr}{A quoted list of logical expressions operating on pattern-derived
       masks which will further combine the masks}
}
\value{
A list (of length \code{length(patterns)+length(expr)}) of
        vectors (each of length \code{length(x)}).
}
\description{
Takes strings, check if they match patterns, and construct expressions according to a string.
}
\examples{
n <- 10
widgets <- data.frame(id=1:n, weight_dry=exp(rnorm(n)),
                      weight_wet=exp(rnorm(n))+10, volume=exp(rnorm(n)))
patterns <- list(weights="^weight")
expr <- quote(list(measurement = weights | volume))
masks <- string_masks(x=names(widgets), patterns=patterns, expr=expr)
}

