% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[,stan_commander,character-method}
\alias{[,stan_commander,character-method}
\title{Accessor to get a vector of samples for a parameter.}
\usage{
\S4method{[}{stan_commander,character}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{stan_commander object}

\item{i}{name of the parameter}

\item{j}{first index, if any.}

\item{...}{additional indexing, if any.}

\item{drop}{currently ignored.}
}
\description{
Accessor to get a vector of samples for a parameter.
}
\details{
Does some checking, then delegates to a stan_commander member
}

