% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trim_whitespace}
\alias{trim_whitespace}
\title{Trims whitespace from a string front and back.}
\usage{
trim_whitespace(s)
}
\arguments{
\item{s}{the string}
}
\value{
s without leading and trailing whitespace.
}
\description{
Trims whitespace from a string front and back.
}

