% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan-names.R
\name{generate_dimensions}
\alias{generate_dimensions}
\title{Use Stan .csv column names to calculate dimension attributes of each
parameter.}
\usage{
generate_dimensions(columns)
}
\arguments{
\item{columns}{.csv column names with name.i.j.... format.}
}
\value{
a named list with dim(parameter) for each element.
}
\description{
Use Stan .csv column names to calculate dimension attributes of each
parameter.
}

