% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grab_imm_diagonal}
\alias{grab_imm_diagonal}
\title{Grab the line of a Stan .csv file with the inverse mass matrix
diagonal in it.}
\usage{
grab_imm_diagonal(lines)
}
\arguments{
\item{lines}{lines from the Stan .csv file, need only enough to get past the recorded meta-data.}
}
\value{
vector of values of diagonal.
}
\description{
Grab the line of a Stan .csv file with the inverse mass matrix
diagonal in it.
}

