% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trim_whitespace}
\alias{trim_whitespace}
\title{Takes names from Stan .csv files and produces an object with parsed
info from comments as well as parameter/transformed parameter names,
a matched R object class, and object dimensions.}
\usage{
trim_whitespace(s)
}
\arguments{
\item{A}{vector of length 1, containing the file name of a Stan .csv output file.}
}
\description{
Takes names from Stan .csv files and produces an object with parsed
info from comments as well as parameter/transformed parameter names,
a matched R object class, and object dimensions.
}
\examples{
parameters <- parse_parameters(file='/tmp/stan.csv')
}

