% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan-names.R
\name{trim_whitespace}
\alias{trim_whitespace}
\title{Trims whitespace from a string front and back.}
\usage{
trim_whitespace(s)
}
\arguments{
\item{s}{the string}
}
\value{
s without leading and trailing whitespace.
}
\description{
Trims whitespace from a string front and back.
}

